#297 [refactor]: JourneyType 하나로 통일하도록 리팩토링
