+++ app/src/main/java/com/byeboo/app/domain/usecase/LoginUseCase.kt
@@ -2,7 +2,6 @@ package com.byeboo.app.domain.usecase
 
 import com.byeboo.app.core.model.auth.TokenEntity
 import com.byeboo.app.domain.model.auth.AuthResult
-import com.byeboo.app.domain.model.auth.toJourneyText
 import com.byeboo.app.domain.repository.auth.AuthRepository
 import com.byeboo.app.domain.repository.auth.TokenRepository
 import com.byeboo.app.domain.repository.auth.UserRepository
@@ -34,7 +33,7 @@ class LoginUseCase
                     )
 
                     auth.name?.let { userRepository.updateUserNickname(it) }
-                    questStateRepository.updateUserJourney(auth.journey.toJourneyText())
+                    questStateRepository.updateUserJourney(auth.journey.journeyName)
                     userRepository.setUserRegistered(auth.isRegistered)
                     userRepository.saveUserId(auth.userId)
